+++ src/angular/hq/src/app/psr/psrtime-list/psrtime-list.component.ts
@@ -46,6 +46,7 @@ import {
 import { FormsModule } from '@angular/forms';
 import { PsrService } from '../psr-service';
 import { ButtonState } from '../../enums/ButtonState';
+import { ModalService } from '../../services/modal.service';
 
 export interface ChargeCodeViewModel {
   id: string;
@@ -106,7 +107,8 @@ export class PSRTimeListComponent implements OnInit, OnDestroy {
     private route: ActivatedRoute,
     private psrService: PsrService,
     private hqSnackBarService: HQSnackBarService,
-    private hqConfirmationModalService: HQConfirmationModalService
+    private hqConfirmationModalService: HQConfirmationModalService,
+    private modalService: ModalService
   ) {
     this.sortOption$ = new BehaviorSubject<SortColumn>(SortColumn.Date);
     this.sortDirection$ = new BehaviorSubject<SortDirection>(SortDirection.Asc);
@@ -321,7 +323,7 @@ export class PSRTimeListComponent implements OnInit, OnDestroy {
     );
 
     if (!time || description.length < 1) {
-      alert('Please Enter a description');
+      this.modalService.alert('Error', 'Please Enter a description');
       // TODO: Alert the users
       return;
     }
@@ -395,7 +397,7 @@ export class PSRTimeListComponent implements OnInit, OnDestroy {
 
     if (!time || !chargeCode || chargeCode.length != 5) {
       // this condition is to check if the charge code is valid
-      alert('Please Enter a Activity Name');
+      this.modalService.alert('Error', 'Please Enter a Activity Name');
       // TODO: Alert the users
       return;
     }
@@ -409,7 +411,8 @@ export class PSRTimeListComponent implements OnInit, OnDestroy {
     //   this.refresh$.next();
     //   return;
     // }
-    const changeChargeCode = window.confirm(
+    const changeChargeCode = this.modalService.confirm(
+      'Confirmation',
       `Are you sure you want to change the charge code to ${chargeCode}?`
     );
     if (!changeChargeCode) {
@@ -449,7 +452,7 @@ export class PSRTimeListComponent implements OnInit, OnDestroy {
     }
 
     if (!time || billableHours == '0' || billableHours == '') {
-      alert('Please Add a time to your billable hours');
+      this.modalService.alert('Error', 'Please Add a time to your billable hours');
       return;
     }
     const chargecodeId = await firstValueFrom(
@@ -485,7 +488,7 @@ export class PSRTimeListComponent implements OnInit, OnDestroy {
     // const actionTaken = await firstValueFrom(
     //   this.hqConfirmationModalService.cuurentAction
     // );
-    const notes = window.prompt('Enter Notes');
+    const notes = await firstValueFrom(this.modalService.prompt('Enter Notes'));
     console.log(notes);
     // if (actionTaken != true) {
     //   this.refresh$.next();