+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Modal service with custom alert, confirm and prompt modals
+
 ## [0.3.0] - 2024-06-09
 
 ### Fixed